fixed imports and fixed an inconsistency in numeric type for numeric constraints

+++ src/main/java/org.metadatacenter.nih.ingestor/converter/CDEConstraints.java
@@ -144,7 +144,7 @@ public class CDEConstraints {
 
         public CDEConstraints build() {
             NumericType numericType = numericPrecision.isPresent() ?
-                    NumericType.DOUBLE : NumericType.INTEGER;
+                    NumericType.DECIMAL : NumericType.INTEGER;
             return new CDEConstraints(datePrecision, permissibleValues, minLength, maxLength,
                     minValue, maxValue, numericPrecision, Optional.of(numericType));
         }
